+++ components/Header/index.tsx
@@ -41,13 +41,8 @@ type MenuItem = {
 const items: MenuItem[] = [
   //{ key: 'home', name: 'Hjem', to: '/' },
   { key: 'program', name: 'Program', to: '/program' },
+  { key: 'stands', name: 'Stands', to: '/stands' },
   { key: 'joblistings', name: 'Jobb', to: '/jobb' },
-  {
-    key: 'info',
-    name: 'Info',
-    to: '/info/[side]',
-    as: '/info/bankett',
-  },
   { key: 'galleri', name: 'Galleri', to: '/galleri' },
   { key: 'faq', name: 'FAQ', to: '/faq' },
   { key: 'about-us', name: 'Om oss', to: '/om-itdagene' },
